Move getOrCreateChain to CifParser
